+++ dash_pydantic_form/fields/all_fields.py
@@ -1,3 +1,5 @@
+import logging
+
 from .base_fields import CheckboxField as Checkbox
 from .base_fields import ChecklistField as Checklist
 from .base_fields import ColorField as Color
@@ -17,13 +19,28 @@ from .base_fields import TextareaField as Textarea
 from .base_fields import TextField as Text
 from .base_fields import TimeField as Time
 from .dict_field import DictField as Dict
-from .editabletable_field import EditableTableField as EditableTable
 from .list_field import ListField as List
 from .markdown_field import MarkdownField as Markdown
 from .model_field import ModelField as Model
+from .table_field import TableField as Table
+
+
+def deprecated_field_factory(name: str, base_class: type):
+    """Create a field class with a deprecation warning message."""
+    old_post_init = getattr(base_class, "model_post_init", None)
+
+    def post_init(self, _context):
+        logging.warning(f"{name} is deprecated, use {base_class.__name__.removesuffix('Field')} instead.")
+        if old_post_init is not None:
+            old_post_init(self, _context)
+
+    newclass = type(name, (base_class,), {"model_post_init": post_init})
+    return newclass
+
 
 # Backwards compatibility
-ModelList = List
+ModelList = deprecated_field_factory("ModelList", List)
+EditableTable = deprecated_field_factory("EditableTable", Table)
 
 __all__ = [
     "Checkbox",
@@ -47,6 +64,7 @@ __all__ = [
     "Select",
     "Slider",
     "Switch",
+    "Table",
     "Textarea",
     "Text",
     "Time",